# bpo-46090: Allow `PyThreadState.datastack_*` members to be `NULL`