<!-- issue-number: [bpo-46090](https://bugs.python.org/issue46090) -->
https://bugs.python.org/issue46090
<!-- /issue-number -->