@brandtbucher 2021-12-23
Perf is good:

```
Slower (16):
- pickle: 9.90 us +- 0.13 us -> 10.5 us +- 0.9 us: 1.06x slower
- pickle_list: 4.38 us +- 0.07 us -> 4.55 us +- 0.06 us: 1.04x slower
- pickle_dict: 26.9 us +- 0.1 us -> 27.9 us +- 0.1 us: 1.04x slower
- telco: 6.42 ms +- 0.11 ms -> 6.65 ms +- 0.11 ms: 1.04x slower
- chameleon: 7.39 ms +- 0.05 ms -> 7.59 ms +- 0.11 ms: 1.03x slower
- unpickle_list: 5.06 us +- 0.05 us -> 5.18 us +- 0.07 us: 1.02x slower
- unpickle_pure_python: 246 us +- 2 us -> 249 us +- 4 us: 1.01x slower
- pathlib: 19.1 ms +- 0.3 ms -> 19.4 ms +- 0.3 ms: 1.01x slower
- json_dumps: 12.7 ms +- 0.1 ms -> 12.8 ms +- 0.2 ms: 1.01x slower
- regex_v8: 23.9 ms +- 0.2 ms -> 24.1 ms +- 0.3 ms: 1.01x slower
- pyflate: 463 ms +- 3 ms -> 466 ms +- 5 ms: 1.01x slower
- json_loads: 25.5 us +- 0.2 us -> 25.7 us +- 0.2 us: 1.01x slower
- sqlite_synth: 2.72 us +- 0.04 us -> 2.73 us +- 0.04 us: 1.01x slower
- sympy_sum: 170 ms +- 2 ms -> 170 ms +- 2 ms: 1.00x slower
- chaos: 74.3 ms +- 0.8 ms -> 74.7 ms +- 0.9 ms: 1.00x slower
- xml_etree_process: 57.3 ms +- 0.5 ms -> 57.5 ms +- 0.5 ms: 1.00x slower

Faster (25):
- pidigits: 202 ms +- 0 ms -> 194 ms +- 0 ms: 1.04x faster
- nqueens: 86.0 ms +- 1.0 ms -> 83.7 ms +- 1.0 ms: 1.03x faster
- deltablue: 4.23 ms +- 0.03 ms -> 4.12 ms +- 0.04 ms: 1.03x faster
- logging_format: 6.52 us +- 0.08 us -> 6.37 us +- 0.07 us: 1.02x faster
- logging_silent: 107 ns +- 1 ns -> 105 ns +- 1 ns: 1.02x faster
- scimark_monte_carlo: 73.8 ms +- 1.3 ms -> 72.2 ms +- 0.8 ms: 1.02x faster
- logging_simple: 5.95 us +- 0.08 us -> 5.83 us +- 0.07 us: 1.02x faster
- scimark_sor: 127 ms +- 1 ms -> 124 ms +- 2 ms: 1.02x faster
- raytrace: 320 ms +- 4 ms -> 314 ms +- 3 ms: 1.02x faster
- mako: 11.5 ms +- 0.2 ms -> 11.3 ms +- 0.1 ms: 1.02x faster
- scimark_fft: 345 ms +- 4 ms -> 340 ms +- 3 ms: 1.02x faster
- richards: 53.5 ms +- 0.7 ms -> 52.8 ms +- 0.8 ms: 1.01x faster
- float: 78.8 ms +- 1.2 ms -> 77.8 ms +- 1.0 ms: 1.01x faster
- hexiom: 6.64 ms +- 0.04 ms -> 6.57 ms +- 0.04 ms: 1.01x faster
- nbody: 99.0 ms +- 1.8 ms -> 97.9 ms +- 1.8 ms: 1.01x faster
- scimark_lu: 113 ms +- 3 ms -> 112 ms +- 2 ms: 1.01x faster
- go: 149 ms +- 3 ms -> 148 ms +- 2 ms: 1.01x faster
- django_template: 35.7 ms +- 0.3 ms -> 35.4 ms +- 0.4 ms: 1.01x faster
- fannkuch: 392 ms +- 5 ms -> 389 ms +- 2 ms: 1.01x faster
- crypto_pyaes: 87.7 ms +- 0.7 ms -> 87.0 ms +- 0.6 ms: 1.01x faster
- python_startup_no_site: 5.81 ms +- 0.00 ms -> 5.78 ms +- 0.00 ms: 1.00x faster
- python_startup: 14.2 ms +- 0.0 ms -> 14.1 ms +- 0.0 ms: 1.00x faster
- regex_compile: 140 ms +- 1 ms -> 139 ms +- 1 ms: 1.00x faster
- 2to3: 266 ms +- 1 ms -> 265 ms +- 1 ms: 1.00x faster
- regex_dna: 216 ms +- 1 ms -> 216 ms +- 1 ms: 1.00x faster

Benchmark hidden because not significant (17): dulwich_log, meteor_contest, pickle_pure_python, regex_effbot, scimark_sparse_mat_mult, spectral_norm, sqlalchemy_declarative, sqlalchemy_imperative, sympy_expand, sympy_integrate, sympy_str, tornado_http, unpack_sequence, unpickle, xml_etree_parse, xml_etree_iterparse, xml_etree_generate

Geometric mean: 1.00x faster
```